Refactored ethereum-icm-verification flows. The two e2e tests have become one
