+++ icm-contracts/tests/flows/ethereum_icm_verification/avalanche_validator_set_registry.go
@@ -3,16 +3,14 @@ package ethereum_icm_verification
 import (
 	"context"
 	"crypto/ecdsa"
-	"crypto/rand"
 	"math/big"
 
-	"github.com/ava-labs/avalanchego/vms/platformvm"
-	adapter "github.com/ava-labs/icm-services/abi-bindings/go/Adapter"
-	diffupdater "github.com/ava-labs/icm-services/abi-bindings/go/DiffUpdater"
+	"github.com/ava-labs/avalanchego/ids"
+	teleportermessengerv2 "github.com/ava-labs/icm-services/abi-bindings/go/TeleporterMessengerV2"
 	ecdsaverifier "github.com/ava-labs/icm-services/abi-bindings/go/mocks/ECDSAVerifier"
 	localnetwork "github.com/ava-labs/icm-services/icm-contracts/tests/network"
 	"github.com/ava-labs/icm-services/icm-contracts/tests/utils"
-	deploymentUtils "github.com/ava-labs/icm-services/icm-contracts/utils/deployment-utils"
+	"github.com/ava-labs/libevm/accounts/abi"
 	"github.com/ava-labs/libevm/accounts/abi/bind"
 	"github.com/ava-labs/libevm/common"
 	"github.com/ava-labs/libevm/crypto"
@@ -21,204 +19,29 @@ import (
 
 // AvalancheValidatorSetRegistry Test that we can deploy a DiffUpdater contract on Ethereum and
 // populate it with the validator set from the Avalanche network.
-// 1. Deploy a DiffUpdater contract on Ethereum
-// 2. Apply the shards to initialize the initialize validator set
+//  1. Deploy the TeleporterMessengerV2 contracts with the ECDSAVerifier/DiffUpdater adapter
+//  2. Send a cross-chain message from Ethereum to Avalanche
+//  3. Manually relay the message: Recover the emitted event and sign it  and submit the signed
+//     message to the contract on Avalanche
 func AvalancheValidatorSetRegistry(
 	ctx context.Context,
 	localAvalancheNetwork *localnetwork.LocalAvalancheNetwork,
 	localEthereumNetwork *localnetwork.LocalEthereumNetwork,
-	ecdsaVerifierByteCodeFile string,
-	adapterByteCodeFile string,
+	ecdsaSigner *ecdsa.PrivateKey,
+	ecdsaVerifierContractAddress common.Address,
+	adapterContractAddress common.Address,
 	teleporterInfo utils.TeleporterTestInfo,
 ) {
 	// set top-level variables
 	_, fundedEthereumKey := localEthereumNetwork.GetFundedAccountInfo()
 	primaryNetworkInfo := localAvalancheNetwork.GetPrimaryNetworkInfo()
 	_, fundedAvalancheKey := localAvalancheNetwork.GetFundedAccountInfo()
-	ethereumNetworkInfo := localEthereumNetwork.GetNetworkInfo()[0]
-	// Get a private key to sign messages from Ethereum
-	ecdsaSigner, err := ecdsa.GenerateKey(crypto.S256(), rand.Reader)
-	Expect(err).Should(BeNil())
+	ethereumNetworkInfo := localEthereumNetwork.EthereumTestInfo()
+	ethereumBlockchainID := localEthereumNetwork.EthereumTestInfo().ChainID()
 
 	// =========================================================================
-	// Step 1: Deploy the DiffUpdater contract on both chains
+	// Step 1: Deploy the TeleporterMessengerV2 contract on both chains
 	// =========================================================================
-
-	// first deploy the DiffUpdater contract on the Ethereum network and initialize it
-	registryContractAddress, serializedShards := utils.DeployDiffUpdater(
-		ctx,
-		localEthereumNetwork.EthereumTestInfo(),
-		fundedEthereumKey,
-		1,
-		primaryNetworkInfo.BlockchainID,
-		primaryNetworkInfo.SubnetID,
-		platformvm.NewClient(primaryNetworkInfo.NodeURIs[0]),
-		5,
-	)
-	// sanity check
-	Expect(len(serializedShards)).Should(Equal(4))
-
-	avalancheValidatorSetRegistry, err := diffupdater.NewDiffUpdater(
-		registryContractAddress,
-		localEthereumNetwork.EthClient,
-	)
-	Expect(err).Should(BeNil())
-
-	opts, err := bind.NewKeyedTransactorWithChainID(fundedEthereumKey, localEthereumNetwork.ChainID)
-	Expect(err).Should(BeNil())
-
-	// apply the shards to initialize the validator set
-	for i, shardBytes := range serializedShards {
-		shard := diffupdater.ValidatorSetShard{
-			AvalancheBlockchainID: primaryNetworkInfo.BlockchainID,
-			ShardNumber:           uint64(i) + 1,
-		}
-		tx, err := avalancheValidatorSetRegistry.UpdateValidatorSet(opts, shard, shardBytes)
-		Expect(err).Should(BeNil())
-		receipt := utils.WaitForTransactionSuccess(ctx, localEthereumNetwork.EthClient, tx.Hash())
-		if i+1 == len(serializedShards) {
-			event, err := utils.GetEventFromLogs(receipt.Logs, avalancheValidatorSetRegistry.ParseValidatorSetUpdated)
-			Expect(err).Should(BeNil())
-			Expect(event.AvalancheBlockchainID).Should(Equal([32]byte(primaryNetworkInfo.BlockchainID)))
-		}
-	}
-	registered, err := avalancheValidatorSetRegistry.IsRegistered(&bind.CallOpts{}, primaryNetworkInfo.BlockchainID)
-	Expect(err).Should(BeNil())
-	Expect(registered).Should(BeTrue())
-
-	// Deploy the DiffUpdater contract on Avalanche
-	// N.B. We don't need to initialize the contract as it will only be used for sending messages
-	contractAddress, _ := utils.DeployDiffUpdater(
-		ctx,
-		&primaryNetworkInfo,
-		fundedAvalancheKey,
-		1,
-		primaryNetworkInfo.BlockchainID,
-		primaryNetworkInfo.SubnetID,
-		platformvm.NewClient(primaryNetworkInfo.NodeURIs[0]),
-		// N.B. This must be the same as above so that the constructor arguments match
-		// for both deployments
-		5,
-	)
-	// Ensure that the contract address is the same as the one deployed on Ethereum
-	Expect(contractAddress).Should(Equal(registryContractAddress))
-
-	// =========================================================================
-	// Step 2: Deploy the ECDSA verifier contract on both chains
-	// =========================================================================
-
-	byteCode, err := deploymentUtils.ExtractByteCodeFromFile(ecdsaVerifierByteCodeFile)
-	Expect(err).Should(BeNil())
-
-	// Generate the ECDSAVerifier deployer transaction via Nick's method
-	// N.B. Use a different gas price to generate a unique deployer address
-	// 		This only shows up as an issue because we re-use Avalanche networks across test
-	ecdsaGasPrice := big.NewInt(0).Add(deploymentUtils.GetDefaultContractCreationGasPrice(), big.NewInt(1))
-	ecdsaVerifierContractTransaction,
-		ecdsaVerifierDeployerAddress,
-		ecdsaVerifierContractAddress,
-		err := deploymentUtils.ConstructKeylessTransaction(
-		byteCode,
-		nil,
-		ecdsaGasPrice,
-		nil,
-	)
-	Expect(err).Should(BeNil())
-	// Deploy the ECDSAVerifier contract on the C-Chain
-	utils.DeployWithNicksMethod(
-		ctx,
-		&primaryNetworkInfo,
-		ecdsaVerifierContractTransaction,
-		ecdsaVerifierDeployerAddress,
-		ecdsaVerifierContractAddress,
-		fundedAvalancheKey,
-	)
-
-	// Initialize the ECDSAVerifier contract on the C-Chain with the `ecdsaSigner` address
-	avalancheEcdsaVerifier, err := ecdsaverifier.NewECDSAVerifier(
-		ecdsaVerifierContractAddress,
-		primaryNetworkInfo.EthClient,
-	)
-	Expect(err).Should(BeNil())
-	opts, err = bind.NewKeyedTransactorWithChainID(fundedAvalancheKey, primaryNetworkInfo.EVMChainID)
-	Expect(err).Should(BeNil())
-	tx, err := avalancheEcdsaVerifier.Initialize(opts, crypto.PubkeyToAddress(ecdsaSigner.PublicKey))
-	Expect(err).Should(BeNil())
-	// Wait for the transaction to be accepted
-	utils.WaitForTransactionSuccess(ctx, primaryNetworkInfo.EthClient, tx.Hash())
-
-	// Deploy the ECDSAVerifier contract on the Ethereum chain
-	utils.DeployWithNicksMethod(
-		ctx,
-		localEthereumNetwork.EthereumTestInfo(),
-		ecdsaVerifierContractTransaction,
-		ecdsaVerifierDeployerAddress,
-		ecdsaVerifierContractAddress,
-		fundedEthereumKey,
-	)
-	// Initialize the ECDSAVerifier contract on Ethereum with the `ecdsaSigner` address
-	ethereumEcdsaVerifier, err := ecdsaverifier.NewECDSAVerifier(
-		ecdsaVerifierContractAddress,
-		localEthereumNetwork.EthClient,
-	)
-	Expect(err).Should(BeNil())
-	opts, err = bind.NewKeyedTransactorWithChainID(fundedEthereumKey, localEthereumNetwork.ChainID)
-	Expect(err).Should(BeNil())
-	tx, err = ethereumEcdsaVerifier.Initialize(opts, crypto.PubkeyToAddress(ecdsaSigner.PublicKey))
-	Expect(err).Should(BeNil())
-	// Wait for the transaction to be accepted
-	utils.WaitForTransactionSuccess(ctx, localEthereumNetwork.EthClient, tx.Hash())
-
-	// =========================================================================
-	// Step 3: Deploy the Adapter contract on both chains
-	// =========================================================================
-	byteCode, err = deploymentUtils.ExtractByteCodeFromFile(adapterByteCodeFile)
-	Expect(err).Should(BeNil())
-
-	// Generate the Adapter deployer transaction via Nick's method
-	adapterABI, err := adapter.AdapterMetaData.GetAbi()
-	Expect(err).Should(BeNil())
-	byteCode, err = deploymentUtils.AddConstructorArgsToByteCode(
-		adapterABI,
-		byteCode,
-		primaryNetworkInfo.BlockchainID,
-		ethereumNetworkInfo.ChainID(),
-		ecdsaVerifierContractAddress,
-		registryContractAddress,
-	)
-	adapterContractTransaction,
-		adapterDeployerAddress,
-		adapterContractAddress,
-		err := deploymentUtils.ConstructKeylessTransaction(
-		byteCode,
-		nil,
-		deploymentUtils.GetDefaultContractCreationGasPrice(),
-		nil,
-	)
-	Expect(err).Should(BeNil())
-	// Deploy the Adapter contract on the C-Chain
-	utils.DeployWithNicksMethod(
-		ctx,
-		&primaryNetworkInfo,
-		adapterContractTransaction,
-		adapterDeployerAddress,
-		adapterContractAddress,
-		fundedAvalancheKey,
-	)
-	// Deploy the Adapter contract on Ethereum
-	utils.DeployWithNicksMethod(
-		ctx,
-		ethereumNetworkInfo,
-		adapterContractTransaction,
-		adapterDeployerAddress,
-		adapterContractAddress,
-		fundedEthereumKey,
-	)
-
-	// =========================================================================
-	// Step 4: Deploy the TeleporterMessengerV2 contract on both chains
-	// =========================================================================
-
 	var teleporterContractAddress common.Address
 	// Deploy the Teleporter registry contracts to all subnets and the C-Chain.
 	for _, l1 := range localAvalancheNetwork.GetAllL1Infos() {
@@ -228,10 +51,114 @@ func AvalancheValidatorSetRegistry(
 	// Deploy the Teleporter registry contracts to the Ethereum network.
 	teleporterContractAddress = utils.DeployTeleporterV2(
 		ctx,
-		localEthereumNetwork.EthereumTestInfo(),
+		ethereumNetworkInfo,
 		adapterContractAddress,
 		fundedEthereumKey,
 	)
-	teleporterInfo.SetTeleporterV2(teleporterContractAddress, localEthereumNetwork.EthereumTestInfo().ChainID())
+	teleporterInfo.SetTeleporterV2(teleporterContractAddress, ethereumNetworkInfo.ChainID())
 
+	ethereumEcdsaVerifier, err := ecdsaverifier.NewECDSAVerifier(
+		ecdsaVerifierContractAddress,
+		localEthereumNetwork.EthClient,
+	)
+
+	// =========================================================================
+	// Step 2: Send cross-chain message from Ethereum to Avalanche
+	// =========================================================================
+
+	// send a message to the TeleporterMessengerV2 on Ethereum and retrieve the event
+	ethTeleporter := teleporterInfo.TeleporterMessengerV2(ethereumNetworkInfo)
+
+	message := teleportermessengerv2.TeleporterMessageInput{
+		DestinationBlockchainID: primaryNetworkInfo.BlockchainID,
+		DestinationAddress:      common.Address{},
+		FeeInfo: teleportermessengerv2.TeleporterFeeInfo{
+			FeeTokenAddress: common.Address{},
+			Amount:          big.NewInt(0),
+		},
+		RequiredGasLimit:        big.NewInt(10),
+		AllowedRelayerAddresses: []common.Address{},
+		Message:                 []byte("Hello Avalanche!"),
+	}
+	opts, err := bind.NewKeyedTransactorWithChainID(fundedEthereumKey, localEthereumNetwork.ChainID)
+	Expect(err).Should(BeNil())
+	tx, err := ethTeleporter.SendCrossChainMessage(opts, message)
+	Expect(err).Should(BeNil())
+	receipt := utils.WaitForTransactionSuccess(ctx, localEthereumNetwork.EthClient, tx.Hash())
+
+	// get the event from the receipt
+	event, err := utils.GetEventFromLogs(receipt.Logs, ethereumEcdsaVerifier.ParseECDSAVerifierSendMessage)
+	Expect(err).Should(BeNil())
+	// check that the event contains the correct message
+	Expect(event.Message.Message).Should(Equal(message.Message))
+
+	// =========================================================================
+	// Step 3: Manually relay the message
+	// =========================================================================
+
+	// sign the message
+	attestation := signMessage(event.Message, ethereumBlockchainID, ecdsaVerifierContractAddress, ecdsaSigner)
+	msg := teleportermessengerv2.TeleporterICMMessage{
+		Message: teleportermessengerv2.TeleporterMessageV2{
+			MessageNonce:            event.Message.MessageNonce,
+			OriginSenderAddress:     event.Message.OriginSenderAddress,
+			OriginTeleporterAddress: event.Message.OriginTeleporterAddress,
+			DestinationBlockchainID: event.Message.DestinationBlockchainID,
+			DestinationAddress:      event.Message.DestinationAddress,
+			RequiredGasLimit:        event.Message.RequiredGasLimit,
+			AllowedRelayerAddresses: event.Message.AllowedRelayerAddresses,
+			Message:                 event.Message.Message,
+		},
+		SourceNetworkID:    0,
+		SourceBlockchainID: ethereumBlockchainID,
+		Attestation:        attestation,
+	}
+
+	// submit the signed message to the TeleporterMessengerV2 contract on Avalanche for verification
+	avalancheTeleporter := teleporterInfo.TeleporterMessengerV2(&primaryNetworkInfo)
+	opts, err = bind.NewKeyedTransactorWithChainID(fundedAvalancheKey, primaryNetworkInfo.EVMChainID)
+	Expect(err).Should(BeNil())
+	tx, err = avalancheTeleporter.ReceiveCrossChainMessage(opts, msg, common.Address{})
+	Expect(err).Should(BeNil())
+	receipt = utils.WaitForTransactionSuccess(ctx, primaryNetworkInfo.EthClient, tx.Hash())
+	// get the event from the receipt
+	receiptEvent, err := utils.GetEventFromLogs(receipt.Logs, avalancheTeleporter.ParseReceiveCrossChainMessage)
+	Expect(err).Should(BeNil())
+	Expect(receiptEvent.Message.Message).Should(Equal(msg.Message.Message))
+	Expect(true).Should(Equal(true))
+}
+
+// Signs a message expected by the ECDSAVerifier according to the EIP-191 standard
+func signMessage(
+	message ecdsaverifier.TeleporterMessageV2,
+	blockchainID ids.ID,
+	ecdsaVerifierContractAddress common.Address,
+	signer *ecdsa.PrivateKey,
+) []byte {
+	parsed, err := ecdsaverifier.ECDSAVerifierMetaData.GetAbi()
+	Expect(err).Should(BeNil())
+
+	// Find the TeleporterMessageV2 type from the contract ABI
+	// It's used in the sendMessage function
+	sendMessageMethod := parsed.Methods["sendMessage"]
+	teleporterMessageType := sendMessageMethod.Inputs[0].Type
+	bytes32Ty, _ := abi.NewType("bytes32", "bytes32", nil)
+	addressTy, _ := abi.NewType("address", "address", nil)
+
+	arguments := abi.Arguments{
+		{Type: teleporterMessageType},
+		{Type: bytes32Ty},
+		{Type: addressTy},
+	}
+	encoded, err := arguments.Pack(message, blockchainID, ecdsaVerifierContractAddress)
+	Expect(err).Should(BeNil())
+	eip191Prefix := []byte("\x19Ethereum Signed Message:\n32")
+	digest := crypto.Keccak256(append(eip191Prefix, crypto.Keccak256(encoded)...))
+	sig, err := crypto.Sign(digest, signer)
+	Expect(err).Should(BeNil())
+	// Adjust V from 0/1 to 27/28 format
+	if sig[64] < 27 {
+		sig[64] += 27
+	}
+	return sig
 }
